Fix block sort order integrity (data loss prevention)
- Add HeadingUpdate struct for atomic heading+sortOrder writes in reorderAllBlocks

- Fix persistReorderedBlocks to cancel previous persist task (currentPersistTask)

- Fix rebuildDocumentContentStatic to use pure function (no state mutation)

- Fix enforceHierarchyAsync to pass headingUpdates to reorderAllBlocks

- Add pushBlockIds/setContentWithBlockIds to BlockSyncService

- Update JS block-sync-plugin for improved change tracking

- Update design.md for block architecture (10 sections)

+++ final final/Views/Sidebar/OutlineSidebar.swift
@@ -10,12 +10,12 @@ import UniformTypeIdentifiers
 /// Transferable wrapper for drag-and-drop
 struct SectionTransfer: Codable, Transferable {
     let id: String
-    let sortOrder: Int
+    let sortOrder: Double
     let headerLevel: Int
     let isSubtreeDrag: Bool      // True when Option-drag includes descendants
     let childIds: [String]       // Ordered descendant IDs for subtree drag
 
-    init(id: String, sortOrder: Int, headerLevel: Int, isSubtreeDrag: Bool = false, childIds: [String] = []) {
+    init(id: String, sortOrder: Double, headerLevel: Int, isSubtreeDrag: Bool = false, childIds: [String] = []) {
         self.id = id
         self.sortOrder = sortOrder
         self.headerLevel = headerLevel
